courses serializer tests mostly working
