+++ courses/factories.py
@@ -233,6 +233,19 @@ class LearnerProgramRecordShareFactory(DjangoModelFactory):
 
     class Meta:
         model = LearnerProgramRecordShare
+        
+@pytest.fixture()
+def program_with_empty_requirements():
+    program = ProgramFactory.create()
+    ProgramRequirementFactory.add_root(program)
+    root_node = program.requirements_root
+
+    root_node.add_child(
+        node_type=ProgramRequirementNodeType.OPERATOR,
+        operator=ProgramRequirement.Operator.ALL_OF,
+        title="Required Courses",
+    )
+    return program
 
 
 @pytest.fixture
